修改组织机构管理分部各自维护部门组织机构,增加权限认证

+++ app/Http/Controllers/DepartmentController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Department;
 use App\Http\Requests\CompanyRequest;
 use App\Http\Requests\DepartmentRequest;
+use App\User;
 use Illuminate\Http\Request;
 
 class DepartmentController extends Controller
@@ -15,9 +16,23 @@ class DepartmentController extends Controller
      */
     public function index()
     {
+        if (!check_permission('user/departments')) {
+            return _404('无权操作');
+        }
         //获取部门信息
-        $menu = Department::get()->toArray();
-        $list = formatTreeData($menu);
+        if (is_administrator()) {
+            $menu = Department::get()->toArray();
+            $list = formatTreeData($menu);
+        } else {
+            //获取所属分部
+            $company_id = get_user_company_id();
+            //获取分部所有部门信息
+            $menu = Department::where(
+                'company_id', $company_id
+            )->orWhere('id', $company_id)->get()->toArray();
+            $list = formatTreeData($menu, 'id', 'parent_id', headquarters('id'));
+        }
+        set_redirect_url();
         return view('user.department.index', compact('list'));
     }
 
@@ -27,6 +42,9 @@ class DepartmentController extends Controller
      */
     public function create()
     {
+        if (!check_permission('user/departments/create')) {
+            return _404('无权操作');
+        }
         return view('user.department.create');
     }
 
@@ -37,6 +55,9 @@ class DepartmentController extends Controller
      */
     public function store(CompanyRequest $request)
     {
+        if (!check_permission('user/departments/create')) {
+            return _404('无权操作');
+        }
         $menu = new Department();
         $menu->parent_id = headquarters('id');
         $menu->name = $request->name;
@@ -59,19 +80,12 @@ class DepartmentController extends Controller
         }
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-    }
-
     //分部编辑
     public function edit($id)
     {
+        if (!check_permission('user/departments/edit')) {
+            return _404('无权操作');
+        }
         $menu = Department::find($id);
         if ($menu) {
             return view('user.department.edit', compact('menu'));
@@ -83,6 +97,9 @@ class DepartmentController extends Controller
     //分部编辑更新
     public function update(CompanyRequest $request, $id)
     {
+        if (!check_permission('user/departments/edit')) {
+            return _404('无权操作');
+        }
         $menu = Department::find($id);
         if ($menu) {
             $headquarters = headquarters('id');
@@ -114,6 +131,9 @@ class DepartmentController extends Controller
     //分部删除
     public function destroy($id)
     {
+        if (!check_permission('user/departments/destroy')) {
+            return _404('无权操作');
+        }
         //判断是否总部
         if (headquarters('id') == $id) {
             return _404('总部信息不能删除');
@@ -122,7 +142,9 @@ class DepartmentController extends Controller
         if (Department::where('parent_id', $id)->first()) {
             return _404('不能删除，请先删除子部门信息');
         }
-
+        if (User::where('department_id', $id)->first()) {
+            return _404('不能删除，请先删除部门人员信息');
+        }
         if (Department::destroy($id)) {
             return response()->json([
                 'message' => '您操作的数据已被删除', 'data' => null
@@ -136,6 +158,9 @@ class DepartmentController extends Controller
     //新增部门
     public function subCreate(DepartmentRequest $request)
     {
+        if (!check_permission('user/departments/sub/create')) {
+            return _404('无权操作');
+        }
         if ($request->method() == 'POST') {
             $menu = new Department();
             $menu->parent_id = $request->parent_id;
@@ -152,7 +177,7 @@ class DepartmentController extends Controller
             if ($menu->save()) {
                 //权限同步
                 return response()->json([
-                    'message' => '添加成功', 'url' => route('departments.index'),
+                    'message' => '添加成功', 'url' => get_redirect_url(),
                     'data' => $menu->toArray(), 'status' => 'success'
                 ]);
             } else {
@@ -168,7 +193,14 @@ class DepartmentController extends Controller
     //编辑部门
     public function subUpdate(DepartmentRequest $request, $id)
     {
-        $menu = Department::find($id);
+        if (!check_permission('user/departments/sub/create')) {
+            return _404('无权操作');
+        }
+        if (is_administrator()) {
+            $menu = Department::find($id);
+        } else {
+            $menu = Department::where('company_id', get_user_company_id())->find($id);
+        }
         if ($request->method() == 'PUT') {
             if ($menu) {
                 $menu->parent_id = $request->parent_id;
