+++ app/Http/Controllers/UserController.php
@@ -17,6 +17,9 @@ class UserController extends Controller
      */
     public function index(Request $request)
     {
+        if (!check_permission('user/users')) {
+            return _404('无权操作');
+        }
         if ($request->ajax()) {
             $sort_field = $request->input('datatable.sort.field')
                 ? $request->input('datatable.sort.field') : 'id';
@@ -24,16 +27,33 @@ class UserController extends Controller
                 ? $request->input('datatable.sort.sort') : 'desc';
             $prepage = $request->input('datatable.pagination.perpage')
                 ? (int)$request->input('datatable.pagination.perpage') : 20;
-            $user = User::with(['department','roles'])->where(
-                'name', 'like',
-                "%{$request->input('datatable.query.search')}%"
-            )->orderBy(
-                $sort_field
-                , $sort)->paginate(
-                $prepage
-                , ['*']
-                , 'datatable.pagination.page'
-            );
+            if (is_administrator()) {
+                $user = User::with(['department', 'roles'])->where(
+                    'name', 'like',
+                    "%{$request->input('datatable.query.search')}%"
+                )->orderBy(
+                    $sort_field
+                    , $sort)->paginate(
+                    $prepage
+                    , ['*']
+                    , 'datatable.pagination.page'
+                );
+            } else {
+                $user = User::with(['department', 'roles'])->where(
+                    'name', 'like',
+                    "%{$request->input('datatable.query.search')}%"
+                )->where(function ($query) {
+                    //获取用户所属分部所有部门
+                    $query->whereIn('department_id', get_company_deparent(get_user_company_id()));
+                })->orderBy(
+                    $sort_field
+                    , $sort)->paginate(
+                    $prepage
+                    , ['*']
+                    , 'datatable.pagination.page'
+                );
+            }
+
             $meta = [
                 'field' => $sort_field,
                 'sort' => $sort,
@@ -56,17 +76,23 @@ class UserController extends Controller
      */
     public function create()
     {
+        if (!check_permission('user/users/create')) {
+            return _404('无权操作');
+        }
         return view('user.default.create');
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(UserRequest $request)
     {
+        if (!check_permission('user/users/create')) {
+            return _404('无权操作');
+        }
         $user = new User;
         $user->username = $request->username;
         $user->name = $request->name;
@@ -76,16 +102,16 @@ class UserController extends Controller
         $user->sex = $request->sex;
         $user->department_id = $request->department_id ? $request->department_id : 0;
         $user->status = $request->status ? 1 : 0;
-        if ($user->save()){
+        if ($user->save()) {
             //授权角色
-            if($request->role_id){
+            if ($request->role_id) {
                 $user->syncRoles($request->role_id);
             }
             return response()->json([
-                'message'=>'保存成功','status'=>'success',
-                'url'=>route('users.index'),'data'=>$user->toArray()
+                'message' => '保存成功', 'status' => 'success',
+                'url' => route('users.index'), 'data' => $user->toArray()
             ]);
-        }else{
+        } else {
             return _404('保存失败');
         }
     }
@@ -93,15 +119,23 @@ class UserController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
     {
-        $user = User::with(['department','roles'])->find($id);
-        if($user){
-            return view('user.default.edit',compact('user'));
-        }else{
+        if (!check_permission('user/users/edit')) {
+            return _404('无权操作');
+        }
+        if (is_administrator()) {
+            $user = User::with(['department', 'roles'])->find($id);
+        } else {
+            $user = User::with(['department', 'roles'])->whereIn('department_id'
+                , get_company_deparent(get_user_company_id()))->find($id);
+        }
+        if ($user) {
+            return view('user.default.edit', compact('user'));
+        } else {
             return _404('你访问的信息不存在');
         }
     }
@@ -109,17 +143,25 @@ class UserController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  \Illuminate\Http\Request $request
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function update(UserRequest $request, $id)
     {
-        $user = User::find($id);
-        if($user){
+        if (!check_permission('user/users/edit')) {
+            return _404('无权操作');
+        }
+        if (is_administrator()) {
+            $user = User::find($id);
+        } else {
+            $user = User::whereIn('department_id'
+                , get_company_deparent(get_user_company_id()))->find($id);
+        }
+        if ($user) {
             $user->username = $request->username;
             $user->name = $request->name;
-            if($request->password){
+            if ($request->password) {
                 $user->password = $request->password;
             }
             $user->email = $request->email;
@@ -127,23 +169,23 @@ class UserController extends Controller
             $user->sex = $request->sex;
             $user->department_id = $request->department_id ? $request->department_id : 0;
             $user->status = $request->status ? 1 : 0;
-            if ($user->save()){
+            if ($user->save()) {
                 //授权角色
-                if(!is_administrator_user($user->id)){
-                    if($request->role_id){
+                if (!is_administrator_user($user->id) && $user->id !== get_current_login_user_info()) {
+                    if ($request->role_id) {
                         $user->syncRoles($request->role_id);
-                    }else{
+                    } else {
                         $user->roles()->detach();
                     }
                 }
                 return response()->json([
-                    'message'=>'保存成功','status'=>'success',
-                    'url'=>route('users.index'),'data'=>$user->toArray()
+                    'message' => '保存成功', 'status' => 'success',
+                    'url' => route('users.index'), 'data' => $user->toArray()
                 ]);
-            }else{
+            } else {
                 return _404('保存失败');
             }
-        }else{
+        } else {
             return _404('您访问信息不存在');
         }
     }
@@ -151,84 +193,117 @@ class UserController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
-        $user = User::find($id);
-        if($user){
-            if(is_administrator_user($user->id)){
-                return  _404('系统管理员不允许被删除');
+        if (!check_permission('user/users/destroy')) {
+            return _404('无权操作');
+        }
+        if (is_administrator()) {
+            $user = User::find($id);
+        } else {
+            $user = User::whereIn('department_id'
+                , get_company_deparent(get_user_company_id()))->find($id);
+        }
+        if ($user) {
+            if ($user->id == get_current_login_user_info()) {
+                return _404('您不能删除自己');
+            }
+            if (is_administrator_user($user->id)) {
+                return _404('系统管理员不允许被删除');
             }
             //判断是否有相关项目
 
-            if($user->delete()){
+            if ($user->delete()) {
                 return response()->json([
-                    'message'=>'您操作的信息已被删除','data'=>$user->toArray(),
-                    'url'=>route('users.index'),'status'=>'success'
+                    'message' => '您操作的信息已被删除', 'data' => $user->toArray(),
+                    'url' => route('users.index'), 'status' => 'success'
                 ]);
             }
             return _404('您操作的数据删除失败，未知错误');
-        }else{
+        } else {
             return _404('您删除的信息不存在');
         }
     }
 
-    public function power(Request $request){
-        Validator::make($request->all(),[
-            'id'=>'bail|required'
-        ],[
-            'id.required'=>'请选择要操作的数据'
+    public function power(Request $request)
+    {
+        if (!check_permission('user/users/power')) {
+            return _404('无权操作');
+        }
+        Validator::make($request->all(), [
+            'id' => 'bail|required'
+        ], [
+            'id.required' => '请选择要操作的数据'
         ])->validate();
         //获取用户实例
-        $user = User::whereIn('id',$request->id)->get();
-        if($user->isNotEmpty()){
-            foreach ($user as $k=>$v){
-                if($v->id != config('auth.administrator')){
-                    if($request->roles){
+        if(is_administrator()){
+            $user = User::whereIn('id', $request->id)->get();
+        }else{
+            $user = User::whereIn('id', $request->id)->whereIn('department_id'
+                , get_company_deparent(get_user_company_id()))->get();
+        }
+        if ($user->isNotEmpty()) {
+            foreach ($user as $k => $v) {
+                if (!is_administrator_user($v->id)
+                    && $v->id !== get_current_login_user_info()) {
+                    if ($request->roles) {
                         $v->syncRoles($request->roles);
-                    }else{
+                    } else {
                         $v->roles()->detach();
                     }
                 }
             }
+        }else{
+            return _404('您操作的数据不存在');
         }
         return response()->json([
-            'message'=>'授权成功','status'=>'success',
-            'data'=>null,'url'=>route('users.index')
+            'message' => '授权成功', 'status' => 'success',
+            'data' => null, 'url' => route('users.index')
         ]);
     }
 
-    public function editPwd(Request $request){
-        Validator::make($request->all(),[
-            'id'=>'bail|required',
-            'password'=>'bail|required|min:6'
-        ],[
-            'id.required'=>'请选择要操作的数据',
-            'password.required'=>'请输入密码',
-            'password.min'=>'密码长度不能小于6位',
+    public function editPwd(Request $request)
+    {
+        if (!check_permission('user/users/edit')) {
+            return _404('无权操作');
+        }
+        Validator::make($request->all(), [
+            'id' => 'bail|required',
+            'password' => 'bail|required|min:6'
+        ], [
+            'id.required' => '请选择要操作的数据',
+            'password.required' => '请输入密码',
+            'password.min' => '密码长度不能小于6位',
         ])->validate();
-
-        $user = User::whereIn('id',$request->id)->get();
-        if($user->isNotEmpty()){
+        if(is_administrator()){
+            $user = User::whereIn('id', $request->id)->get();
+        }else{
+            $user = User::whereIn('id', $request->id)->whereIn('department_id'
+                , get_company_deparent(get_user_company_id()))->get();
+        }
+        if ($user->isNotEmpty()) {
             $update = [];
-            foreach ($user as $k=>$v){
-                if(is_administrator()){
+            foreach ($user as $k => $v) {
+                if (is_administrator()) {
                     $update[] = $v->id;
-                }elseif($v->id != config('auth.administrator')){
+                } elseif (!is_administrator_user($v->id)) {
                     $update[] = $v->id;
                 }
             }
-            if(count($update)){
-                User::whereIn('id',$update)->update([
-                    'password'=>bcrypt($request->password)
+            if (count($update)) {
+                User::whereIn('id', $update)->update([
+                    'password' => bcrypt($request->password)
                 ]);
             }
+        }else{
+           return _404('您操作的数据不存在');
         }
         return response()->json([
-            'message'=>'重置成功，新密码为 '.$request->password,'status'=>'success',
-            'data'=>null,'url'=>route('users.index')
+            'message' => '重置成功，新密码为 ' . $request->password, 'status' => 'success',
+            'data' => null, 'url' => route('users.index')
         ]);
     }
 }
